kc: Rename error.h -> exceptions.h
This conforms with the convention used in other parts of the codebase
