+++ src/v/kafka/client/client.cc
@@ -12,7 +12,7 @@
 #include "kafka/client/broker.h"
 #include "kafka/client/configuration.h"
 #include "kafka/client/consumer.h"
-#include "kafka/client/error.h"
+#include "kafka/client/exceptions.h"
 #include "kafka/client/logger.h"
 #include "kafka/client/retry_with_mitigation.h"
 #include "kafka/protocol/errors.h"
