+++ src/v/kafka/client/consumer.cc
@@ -11,7 +11,7 @@
 
 #include "kafka/client/assignment_plans.h"
 #include "kafka/client/configuration.h"
-#include "kafka/client/error.h"
+#include "kafka/client/exceptions.h"
 #include "kafka/client/logger.h"
 #include "kafka/protocol/describe_groups.h"
 #include "kafka/protocol/errors.h"
