+++ src/v/kafka/client/exceptions.h
